Modifico  eliminarContacto
Modifico el método   eliminarContacto(id:number){}
para que al momento de que le den "eliminar" al contacto les aparezca una alerta diciendo que si están seguros o no de eliminar el contacto

+++ gestion-contactos-frontend/src/app/lista-contactos/lista-contactos.component.ts
@@ -3,11 +3,12 @@ import { Component, OnInit } from '@angular/core';
 import { Contacto } from '../contacto';
 import { ContactoService } from '../contacto.service';
 import { Router } from '@angular/router';
+import swal from 'sweetalert2';
 
 @Component({
   selector: 'app-lista-contactos',
   templateUrl: './lista-contactos.component.html',
-  styleUrl: './lista-contactos.component.css'
+  styleUrls: ['./lista-contactos.component.css']
 })
 export class ListaContactosComponent implements OnInit{
 
@@ -19,12 +20,29 @@ export class ListaContactosComponent implements OnInit{
   actualizarContacto(id:number){
     this.router.navigate(['actualizar-contacto',id]);
   }
-
   eliminarContacto(id:number){
-    this.contactoServicio.eliminarContacto(id).subscribe(dato => {
-      console.log(dato);
-      this.obtenerContactos();
-    });
+    swal.fire({
+      title: '¿Estás seguro?',
+      text: 'Confirma si deseas eliminar al empleado',
+      icon: 'warning',  // Cambiado de 'type' a 'icon'
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí, elimínalo',
+      cancelButtonText: 'No, cancelar'
+    }).then((result) => {
+      if(result.isConfirmed){
+        this.contactoServicio.eliminarContacto(id).subscribe(dato => {
+          console.log(dato);
+          this.obtenerContactos();
+          swal.fire(
+            'Empleado eliminado',
+            'El empleado ha sido eliminado con exito',
+            'success'
+          )
+        })
+      }
+    })
   }
 
   //Este método hace que me suscriba a toda esa lista
